+++ backend/controllers/UserController.php
@@ -111,7 +111,9 @@ class UserController extends Controller
      */
     public function actionDel()
     {
-        User::deleteAll(['in', 'id', Yii::$app->request->post()['id']]);
+        if(isset(Yii::$app->request->post()['id'])) {
+            User::deleteAll(['in', 'id', Yii::$app->request->post()['id']]);
+        }
         return $this->redirect(Yii::$app->urlHelper->to(['user/index']));
     }
